# s2: Simplify mem-move