--- a/s2/_generate/go.mod
+++ b/s2/_generate/go.mod
@@ -1,0 +1,5 @@
+module github.com/klauspost/compress
+
+go 1.13
+
+require github.com/mmcloughlin/avo v0.0.0-20200523190732-4439b6b2c061