… s2/_generate/go.sum
… s2/testdata/enc_regressions.zip