Removes AVX code

Inspired by https://go-review.googlesource.com/c/go/+/228820

Stream before/after
```
enwik10	s2	1	100000000000	47646440409	16918	5636.71
enwik10	s2	1	100000000000	47614860240	16699	5710.62

consensus.db.10gb	s2	1	107374182400	45523066020	11271	9085.26
consensus.db.10gb	s2	1	107374182400	45626485920	11371	9004.57

sofia-air-quality-dataset.tar	s2	1	154644638720	49917736859	20675	7132.94
sofia-air-quality-dataset.tar	s2	1	154644638720	49917619450	20301	7264.34
```

Simple benchmarks are very noisy. It seems there is a regression on predicted moves.

```
benchmark                                                  old ns/op     new ns/op     delta
BenchmarkRandomEncode                                      32509         28147         -13.42%
BenchmarkRandomEncode-16                                   32555         28171         -13.47%
BenchmarkRandomEncodeBetter                                41796         37732         -9.72%
BenchmarkRandomEncodeBetter-16                             41732         37964         -9.03%
BenchmarkEncodeS2Block/0-html/block                        88275         85477         -3.17%
BenchmarkEncodeS2Block/0-html/block-16                     6100          5970          -2.13%
BenchmarkEncodeS2Block/0-html/block-better                 180681        177768        -1.61%
BenchmarkEncodeS2Block/0-html/block-better-16              13607         13850         +1.79%
BenchmarkEncodeS2Block/1-urls/block                        1108411       1031523       -6.94%
BenchmarkEncodeS2Block/1-urls/block-16                     77007         75679         -1.72%
BenchmarkEncodeS2Block/1-urls/block-better                 2348698       2369262       +0.88%
BenchmarkEncodeS2Block/1-urls/block-better-16              181948        181473        -0.26%
BenchmarkEncodeS2Block/2-jpg/block                         6197          5463          -11.84%
BenchmarkEncodeS2Block/2-jpg/block-16                      459           452           -1.53%
BenchmarkEncodeS2Block/2-jpg/block-better                  13907         13174         -5.27%
BenchmarkEncodeS2Block/2-jpg/block-better-16               1113          1123          +0.90%
BenchmarkEncodeS2Block/3-jpg_200b/block                    157           163           +3.82%
BenchmarkEncodeS2Block/3-jpg_200b/block-16                 11.5          19.4          +68.70%
BenchmarkEncodeS2Block/3-jpg_200b/block-better             3201          3152          -1.53%
BenchmarkEncodeS2Block/3-jpg_200b/block-better-16          284           279           -1.76%
BenchmarkEncodeS2Block/4-pdf/block                         14689         14886         +1.34%
BenchmarkEncodeS2Block/4-pdf/block-16                      1028          1063          +3.40%
BenchmarkEncodeS2Block/4-pdf/block-better                  195500        166266        -14.95%
BenchmarkEncodeS2Block/4-pdf/block-better-16               13717         13772         +0.40%
BenchmarkEncodeS2Block/5-html4/block                       116312        113611        -2.32%
BenchmarkEncodeS2Block/5-html4/block-16                    8296          8053          -2.93%
BenchmarkEncodeS2Block/5-html4/block-better                239375        236458        -1.22%
BenchmarkEncodeS2Block/5-html4/block-better-16             18293         18025         -1.47%
BenchmarkEncodeS2Block/6-txt1/block                        371555        351635        -5.36%
BenchmarkEncodeS2Block/6-txt1/block-16                     25250         25232         -0.07%
BenchmarkEncodeS2Block/6-txt1/block-better                 710719        703302        -1.04%
BenchmarkEncodeS2Block/6-txt1/block-better-16              53071         52486         -1.10%
BenchmarkEncodeS2Block/7-txt2/block                        281605        274912        -2.38%
BenchmarkEncodeS2Block/7-txt2/block-16                     20100         19849         -1.25%
BenchmarkEncodeS2Block/7-txt2/block-better                 624867        616596        -1.32%
BenchmarkEncodeS2Block/7-txt2/block-better-16              48301         46692         -3.33%
BenchmarkEncodeS2Block/8-txt3/block                        1056252       1022727       -3.17%
BenchmarkEncodeS2Block/8-txt3/block-16                     73951         71686         -3.06%
BenchmarkEncodeS2Block/8-txt3/block-better                 1903381       1891548       -0.62%
BenchmarkEncodeS2Block/8-txt3/block-better-16              145467        140893        -3.14%
BenchmarkEncodeS2Block/9-txt4/block                        1247918       1180556       -5.40%
BenchmarkEncodeS2Block/9-txt4/block-16                     88036         85297         -3.11%
BenchmarkEncodeS2Block/9-txt4/block-better                 2512764       2497889       -0.59%
BenchmarkEncodeS2Block/9-txt4/block-better-16              191333        188896        -1.27%
BenchmarkEncodeS2Block/10-pb/block                         68251         66888         -2.00%
BenchmarkEncodeS2Block/10-pb/block-16                      4667          4594          -1.56%
BenchmarkEncodeS2Block/10-pb/block-better                  147067        141638        -3.69%
BenchmarkEncodeS2Block/10-pb/block-better-16               10982         10692         -2.64%
BenchmarkEncodeS2Block/11-gaviota/block                    318400        312981        -1.70%
BenchmarkEncodeS2Block/11-gaviota/block-16                 22266         21800         -2.09%
BenchmarkEncodeS2Block/11-gaviota/block-better             560836        557985        -0.51%
BenchmarkEncodeS2Block/11-gaviota/block-better-16          42364         41923         -1.04%
BenchmarkEncodeS2Block/12-txt1_128b/block                  117           121           +3.42%
BenchmarkEncodeS2Block/12-txt1_128b/block-16               8.70          8.83          +1.49%
BenchmarkEncodeS2Block/12-txt1_128b/block-better           2873          2865          -0.28%
BenchmarkEncodeS2Block/12-txt1_128b/block-better-16        248           237           -4.44%
BenchmarkEncodeS2Block/13-txt1_1000b/block                 1073          903           -15.84%
BenchmarkEncodeS2Block/13-txt1_1000b/block-16              81.7          68.8          -15.79%
BenchmarkEncodeS2Block/13-txt1_1000b/block-better          6438          6386          -0.81%
BenchmarkEncodeS2Block/13-txt1_1000b/block-better-16       549           517           -5.83%
BenchmarkEncodeS2Block/14-txt1_10000b/block                11685         10684         -8.57%
BenchmarkEncodeS2Block/14-txt1_10000b/block-16             817           784           -4.04%
BenchmarkEncodeS2Block/14-txt1_10000b/block-better         43519         42859         -1.52%
BenchmarkEncodeS2Block/14-txt1_10000b/block-better-16      3401          3351          -1.47%
BenchmarkEncodeS2Block/15-txt1_20000b/block                24160         26198         +8.44%
BenchmarkEncodeS2Block/15-txt1_20000b/block-16             1767          1747          -1.13%
BenchmarkEncodeS2Block/15-txt1_20000b/block-better         93892         92425         -1.56%
BenchmarkEncodeS2Block/15-txt1_20000b/block-better-16      7456          7218          -3.19%
BenchmarkEncodeSnappyBlock/0-html/s2-snappy                87445         87401         -0.05%
BenchmarkEncodeSnappyBlock/0-html/s2-snappy-16             5942          5968          +0.44%
BenchmarkEncodeSnappyBlock/1-urls/s2-snappy                1099075       1062727       -3.31%
BenchmarkEncodeSnappyBlock/1-urls/s2-snappy-16             75932         75547         -0.51%
BenchmarkEncodeSnappyBlock/2-jpg/s2-snappy                 6670          6363          -4.60%
BenchmarkEncodeSnappyBlock/2-jpg/s2-snappy-16              530           469           -11.51%
BenchmarkEncodeSnappyBlock/3-jpg_200b/s2-snappy            155           152           -1.94%
BenchmarkEncodeSnappyBlock/3-jpg_200b/s2-snappy-16         19.7          19.3          -2.03%
BenchmarkEncodeSnappyBlock/4-pdf/s2-snappy                 14558         14613         +0.38%
BenchmarkEncodeSnappyBlock/4-pdf/s2-snappy-16              1073          1068          -0.47%
BenchmarkEncodeSnappyBlock/5-html4/s2-snappy               110911        108111        -2.52%
BenchmarkEncodeSnappyBlock/5-html4/s2-snappy-16            7808          7618          -2.43%
BenchmarkEncodeSnappyBlock/6-txt1/s2-snappy                372609        356879        -4.22%
BenchmarkEncodeSnappyBlock/6-txt1/s2-snappy-16             25589         25451         -0.54%
BenchmarkEncodeSnappyBlock/7-txt2/s2-snappy                279691        290072        +3.71%
BenchmarkEncodeSnappyBlock/7-txt2/s2-snappy-16             19983         20083         +0.50%
BenchmarkEncodeSnappyBlock/8-txt3/s2-snappy                1029301       1042907       +1.32%
BenchmarkEncodeSnappyBlock/8-txt3/s2-snappy-16             72743         72419         -0.45%
BenchmarkEncodeSnappyBlock/9-txt4/s2-snappy                1216904       1211211       -0.47%
BenchmarkEncodeSnappyBlock/9-txt4/s2-snappy-16             87610         86175         -1.64%
BenchmarkEncodeSnappyBlock/10-pb/s2-snappy                 66515         67217         +1.06%
BenchmarkEncodeSnappyBlock/10-pb/s2-snappy-16              4614          4643          +0.63%
BenchmarkEncodeSnappyBlock/11-gaviota/s2-snappy            311467        313867        +0.77%
BenchmarkEncodeSnappyBlock/11-gaviota/s2-snappy-16         21867         21736         -0.60%
BenchmarkEncodeSnappyBlock/12-txt1_128b/s2-snappy          118           117           -0.85%
BenchmarkEncodeSnappyBlock/12-txt1_128b/s2-snappy-16       18.2          8.83          -51.48%
BenchmarkEncodeSnappyBlock/13-txt1_1000b/s2-snappy         1074          1068          -0.56%
BenchmarkEncodeSnappyBlock/13-txt1_1000b/s2-snappy-16      79.5          85.4          +7.42%
BenchmarkEncodeSnappyBlock/14-txt1_10000b/s2-snappy        10615         10582         -0.31%
BenchmarkEncodeSnappyBlock/14-txt1_10000b/s2-snappy-16     789           787           -0.25%
BenchmarkEncodeSnappyBlock/15-txt1_20000b/s2-snappy        26296         26661         +1.39%
BenchmarkEncodeSnappyBlock/15-txt1_20000b/s2-snappy-16     1758          1789          +1.76%

benchmark                                                  old MB/s      new MB/s      speedup
BenchmarkRandomEncode                                      32255.06      37253.25      1.15x
BenchmarkRandomEncode-16                                   32209.29      37222.27      1.16x
BenchmarkRandomEncodeBetter                                25087.87      27790.32      1.11x
BenchmarkRandomEncodeBetter-16                             25126.46      27620.60      1.10x
BenchmarkEncodeS2Block/0-html/block                        1160.01       1197.98       1.03x
BenchmarkEncodeS2Block/0-html/block-16                     16788.24      17152.43      1.02x
BenchmarkEncodeS2Block/0-html/block-better                 566.74        576.03        1.02x
BenchmarkEncodeS2Block/0-html/block-better-16              7525.78       7393.66       0.98x
BenchmarkEncodeS2Block/1-urls/block                        633.42        680.63        1.07x
BenchmarkEncodeS2Block/1-urls/block-16                     9117.14       9277.15       1.02x
BenchmarkEncodeS2Block/1-urls/block-better                 298.93        296.33        0.99x
BenchmarkEncodeS2Block/1-urls/block-better-16              3858.73       3868.83       1.00x
BenchmarkEncodeS2Block/2-jpg/block                         19864.83      22530.75      1.13x
BenchmarkEncodeS2Block/2-jpg/block-16                      268412.83     272349.31     1.01x
BenchmarkEncodeS2Block/2-jpg/block-better                  8851.23       9343.92       1.06x
BenchmarkEncodeS2Block/2-jpg/block-better-16               110596.41     109610.76     0.99x
BenchmarkEncodeS2Block/3-jpg_200b/block                    1273.34       1229.06       0.97x
BenchmarkEncodeS2Block/3-jpg_200b/block-16                 17436.81      10312.53      0.59x
BenchmarkEncodeS2Block/3-jpg_200b/block-better             62.48         63.46         1.02x
BenchmarkEncodeS2Block/3-jpg_200b/block-better-16          703.62        718.09        1.02x
BenchmarkEncodeS2Block/4-pdf/block                         6971.16       6878.97       0.99x
BenchmarkEncodeS2Block/4-pdf/block-16                      99610.88      96331.07      0.97x
BenchmarkEncodeS2Block/4-pdf/block-better                  523.79        615.88        1.18x
BenchmarkEncodeS2Block/4-pdf/block-better-16               7465.10       7435.29       1.00x
BenchmarkEncodeS2Block/5-html4/block                       3521.57       3605.28       1.02x
BenchmarkEncodeS2Block/5-html4/block-16                    49376.06      50862.24      1.03x
BenchmarkEncodeS2Block/5-html4/block-better                1711.12       1732.23       1.01x
BenchmarkEncodeS2Block/5-html4/block-better-16             22391.00      22723.52      1.01x
BenchmarkEncodeS2Block/6-txt1/block                        409.33        432.52        1.06x
BenchmarkEncodeS2Block/6-txt1/block-16                     6023.33       6027.53       1.00x
BenchmarkEncodeS2Block/6-txt1/block-better                 213.99        216.25        1.01x
BenchmarkEncodeS2Block/6-txt1/block-better-16              2865.74       2897.71       1.01x
BenchmarkEncodeS2Block/7-txt2/block                        444.52        455.34        1.02x
BenchmarkEncodeS2Block/7-txt2/block-16                     6227.92       6306.59       1.01x
BenchmarkEncodeS2Block/7-txt2/block-better                 200.33        203.02        1.01x
BenchmarkEncodeS2Block/7-txt2/block-better-16              2591.62       2680.97       1.03x
BenchmarkEncodeS2Block/8-txt3/block                        404.03        417.27        1.03x
BenchmarkEncodeS2Block/8-txt3/block-16                     5770.76       5953.10       1.03x
BenchmarkEncodeS2Block/8-txt3/block-better                 224.21        225.61        1.01x
BenchmarkEncodeS2Block/8-txt3/block-better-16              2933.68       3028.93       1.03x
BenchmarkEncodeS2Block/9-txt4/block                        386.13        408.16        1.06x
BenchmarkEncodeS2Block/9-txt4/block-16                     5473.45       5649.22       1.03x
BenchmarkEncodeS2Block/9-txt4/block-better                 191.77        192.91        1.01x
BenchmarkEncodeS2Block/9-txt4/block-better-16              2518.44       2550.93       1.01x
BenchmarkEncodeS2Block/10-pb/block                         1737.52       1772.94       1.02x
BenchmarkEncodeS2Block/10-pb/block-16                      25409.23      25812.47      1.02x
BenchmarkEncodeS2Block/10-pb/block-better                  806.36        837.26        1.04x
BenchmarkEncodeS2Block/10-pb/block-better-16               10798.63      11091.27      1.03x
BenchmarkEncodeS2Block/11-gaviota/block                    578.89        588.92        1.02x
BenchmarkEncodeS2Block/11-gaviota/block-16                 8277.94       8454.91       1.02x
BenchmarkEncodeS2Block/11-gaviota/block-better             328.65        330.33        1.01x
BenchmarkEncodeS2Block/11-gaviota/block-better-16          4350.82       4396.64       1.01x
BenchmarkEncodeS2Block/12-txt1_128b/block                  1097.55       1057.39       0.96x
BenchmarkEncodeS2Block/12-txt1_128b/block-16               14708.73      14494.80      0.99x
BenchmarkEncodeS2Block/12-txt1_128b/block-better           44.56         44.68         1.00x
BenchmarkEncodeS2Block/12-txt1_128b/block-better-16        516.96        539.46        1.04x
BenchmarkEncodeS2Block/13-txt1_1000b/block                 931.89        1107.78       1.19x
BenchmarkEncodeS2Block/13-txt1_1000b/block-16              12232.80      14534.00      1.19x
BenchmarkEncodeS2Block/13-txt1_1000b/block-better          155.34        156.60        1.01x
BenchmarkEncodeS2Block/13-txt1_1000b/block-better-16       1821.65       1935.05       1.06x
BenchmarkEncodeS2Block/14-txt1_10000b/block                855.77        936.00        1.09x
BenchmarkEncodeS2Block/14-txt1_10000b/block-16             12239.33      12753.89      1.04x
BenchmarkEncodeS2Block/14-txt1_10000b/block-better         229.78        233.32        1.02x
BenchmarkEncodeS2Block/14-txt1_10000b/block-better-16      2940.43       2983.93       1.01x
BenchmarkEncodeS2Block/15-txt1_20000b/block                827.82        763.43        0.92x
BenchmarkEncodeS2Block/15-txt1_20000b/block-16             11319.00      11445.16      1.01x
BenchmarkEncodeS2Block/15-txt1_20000b/block-better         213.01        216.39        1.02x
BenchmarkEncodeS2Block/15-txt1_20000b/block-better-16      2682.28       2770.72       1.03x
BenchmarkEncodeSnappyBlock/0-html/s2-snappy                1171.02       1171.61       1.00x
BenchmarkEncodeSnappyBlock/0-html/s2-snappy-16             17234.25      17158.71      1.00x
BenchmarkEncodeSnappyBlock/1-urls/s2-snappy                638.80        660.65        1.03x
BenchmarkEncodeSnappyBlock/1-urls/s2-snappy-16             9246.29       9293.34       1.01x
BenchmarkEncodeSnappyBlock/2-jpg/s2-snappy                 18454.82      19344.90      1.05x
BenchmarkEncodeSnappyBlock/2-jpg/s2-snappy-16              232075.96     262204.81     1.13x
BenchmarkEncodeSnappyBlock/3-jpg_200b/s2-snappy            1294.38       1311.82       1.01x
BenchmarkEncodeSnappyBlock/3-jpg_200b/s2-snappy-16         10127.69      10374.78      1.02x
BenchmarkEncodeSnappyBlock/4-pdf/s2-snappy                 7033.78       7007.28       1.00x
BenchmarkEncodeSnappyBlock/4-pdf/s2-snappy-16              95433.22      95880.16      1.00x
BenchmarkEncodeSnappyBlock/5-html4/s2-snappy               3693.05       3788.71       1.03x
BenchmarkEncodeSnappyBlock/5-html4/s2-snappy-16            52459.88      53766.66      1.02x
BenchmarkEncodeSnappyBlock/6-txt1/s2-snappy                408.17        426.16        1.04x
BenchmarkEncodeSnappyBlock/6-txt1/s2-snappy-16             5943.57       5975.71       1.01x
BenchmarkEncodeSnappyBlock/7-txt2/s2-snappy                447.56        431.54        0.96x
BenchmarkEncodeSnappyBlock/7-txt2/s2-snappy-16             6264.25       6233.21       1.00x
BenchmarkEncodeSnappyBlock/8-txt3/s2-snappy                414.61        409.20        0.99x
BenchmarkEncodeSnappyBlock/8-txt3/s2-snappy-16             5866.61       5892.87       1.00x
BenchmarkEncodeSnappyBlock/9-txt4/s2-snappy                395.97        397.83        1.00x
BenchmarkEncodeSnappyBlock/9-txt4/s2-snappy-16             5500.07       5591.67       1.02x
BenchmarkEncodeSnappyBlock/10-pb/s2-snappy                 1782.87       1764.27       0.99x
BenchmarkEncodeSnappyBlock/10-pb/s2-snappy-16              25700.44      25542.43      0.99x
BenchmarkEncodeSnappyBlock/11-gaviota/s2-snappy            591.78        587.26        0.99x
BenchmarkEncodeSnappyBlock/11-gaviota/s2-snappy-16         8429.13       8480.00       1.01x
BenchmarkEncodeSnappyBlock/12-txt1_128b/s2-snappy          1088.43       1094.02       1.01x
BenchmarkEncodeSnappyBlock/12-txt1_128b/s2-snappy-16       7047.17       14489.73      2.06x
BenchmarkEncodeSnappyBlock/13-txt1_1000b/s2-snappy         931.10        936.33        1.01x
BenchmarkEncodeSnappyBlock/13-txt1_1000b/s2-snappy-16      12572.00      11709.65      0.93x
BenchmarkEncodeSnappyBlock/14-txt1_10000b/s2-snappy        942.06        944.98        1.00x
BenchmarkEncodeSnappyBlock/14-txt1_10000b/s2-snappy-16     12670.62      12711.81      1.00x
BenchmarkEncodeSnappyBlock/15-txt1_20000b/s2-snappy        760.58        750.15        0.99x
BenchmarkEncodeSnappyBlock/15-txt1_20000b/s2-snappy-16     11377.66      11178.20      0.98x 
```